beacon/backend/requirements.txt
django==4.2
djangorestframework==3.15
djangorestframework-simplejwt==5.3
django-cors-headers==4.3
psycopg2-binary==2.9
neomodel==5.2
pinecone-client==3.0
openai==1.0
celery==5.3
redis==5.0
python-dotenv==1.0
